Added tile for Jonas' Mesos+Docker+Rexray Vagrant setup
the jQuery tooltip function to load information will not work because the jQuery is examining the API of the repo from the link provided. In this case, we are providing a link to a particular branch. Until Jonas merges it all into the master branch, the tooltip will never work for this particular tile.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>