Small Tweaks
This commit fixed a bunch of a small issues. Renamed tags with Github to GitHub. Removed Focus Communities of GitHub, Coveralls, ReadTheDocs, and Slack. This is repeated because there is a section within the stick that says 'Join our community' that has both Slack and GitHub links. Updated ECS Service Broker for CF titles and edited the tags associated with it. Removed the Spring tag that was from a prior commit. Removed a bunch of old code that had been commented out for a long time and was never going to make it back in.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>