the official cutover
as the sun shines on September 7th 2016, the days of once being known as EMC {code} have come to a close. From hence forth, the group shall be known as {code} by Dell EMC. Today, history is made. We at {code} have been granted an opportunity to make one of the largest tech companies in existence to be a major player in the Open Source Software world. This mission is both challenging and relentless but we take pride in knowing we have crossed many hurdle to make it to this moment. Here's to the future...     also, there's probably going to be some dead links as things transition and move, those will be fixed as we go.

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>